% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cross_weights}
\alias{cross_weights}
\title{Compute cross weight matrix}
\usage{
cross_weights(w)
}
\arguments{
\item{w}{A vector containing weights}
}
\value{
Matrix of squared weights
}
\description{
This function returns the doubly centered version of a matrix
}
