% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcenter}
\alias{dcenter}
\title{Compute doubly centered matrix}
\usage{
dcenter(x)
}
\arguments{
\item{x}{A matrix which to doubly center}
}
\value{
Doubly centered version of a matrix
}
\description{
This function returns the doubly centered version of a matrix
}
