% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_sample_size.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Effective sample size}
\usage{
effective_sample_size(weights)
}
\arguments{
\item{weights}{Weights for each data point}
}
\value{
Effective sample size
}
\description{
Effective sample size
}
