% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_correlation.R
\name{weighted_correlation_table}
\alias{weighted_correlation_table}
\title{Weighted correlations}
\usage{
weighted_correlation_table(treatment, data, weights)
}
\value{
unweighted and weighted correlations
}
\description{
Weighted correlations
}
