% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_dcenter}
\alias{weighted_dcenter}
\title{Compute doubly centered matrix}
\usage{
weighted_dcenter(x, w)
}
\arguments{
\item{x}{A matrix which to doubly center: distance matrix x}

\item{w}{A matrix containing weights}
}
\value{
Doubly centered version of a matrix
}
\description{
This function returns the doubly centered version of a matrix
}
