% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_dcor}
\alias{weighted_dcor}
\title{Compute weighted distance correlation}
\usage{
weighted_dcor(x, y, w)
}
\arguments{
\item{x}{Distance matrix for x}

\item{y}{Distance matrix for y}

\item{w}{Weights}
}
\value{
Distance correlation
}
\description{
This function computes the distance correlation between x and y, where data points are weighted by w
}
