% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_distance_correlation.R
\name{weighted_dcorr_table}
\alias{weighted_dcorr_table}
\title{Table of unweighted and weighted distance correlation}
\usage{
weighted_dcorr_table(treatment, data, weights = rep(1, length(treatment)))
}
\arguments{
\item{treatment}{Treatment variable (binary)}

\item{data}{Data for which rebalancing should be evaluated}

\item{weights}{Propensity weights for each data point}
}
\value{
unweighted and weighted distance correlation
}
\description{
Table of unweighted and weighted distance correlation
}
