% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ks.R
\name{weighted_ecdf}
\alias{weighted_ecdf}
\title{Compute weighted empirical CDF}
\usage{
weighted_ecdf(x, weights = NULL)
}
\arguments{
\item{x}{Empirical data points for which to compute weighted CDF}

\item{weights}{Weights of data points}
}
\value{
Weighted empirical CDF function
}
\description{
Compute weighted empirical CDF
}
