% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ks.R
\name{weighted_ks_stat}
\alias{weighted_ks_stat}
\title{Weighted KS statistic}
\usage{
weighted_ks_stat(treatment, data, weights = rep(1, length(treatment)))
}
\arguments{
\item{treatment}{Binary treatment variable}

\item{data}{Confounders for which to compute KS statistic}

\item{weights}{Weights for each data point}
}
\value{
KS statistic between treated and untreated group for each confounder
}
\description{
Weighted KS statistic
}
