% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ks.R
\name{weighted_ks_table}
\alias{weighted_ks_table}
\title{Table of unweighted and weighted KS statistics}
\usage{
weighted_ks_table(treatment, data, weights)
}
\arguments{
\item{treatment}{Treatment variable (binary)}

\item{data}{Data for which rebalancing should be evaluated}

\item{weights}{Propensity weights for each data point}
}
\value{
unweighted and weighted KS statistics
}
\description{
Table of unweighted and weighted KS statistics
}
