% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Weighted mean}
\usage{
weighted_mean(x, w)
}
\arguments{
\item{x}{A numeric vector for which to compute the weighted mean}

\item{w}{A numeric vector of the same length as x}
}
\value{
Returns the weighted mean
}
\description{
This function returns the weighted mean of a vector
}
