% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_sd}
\alias{weighted_sd}
\title{Weighted sd}
\usage{
weighted_sd(x, w)
}
\arguments{
\item{x}{A numeric vector for which to compute the weighted sd}

\item{w}{A numeric vector of the same length as x}
}
\value{
Returns the weighted sd
}
\description{
This function returns the weighted sd of a vector
}
