% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_smd.R
\name{weighted_smd_stat}
\alias{weighted_smd_stat}
\title{Weighted SMD stat}
\usage{
weighted_smd_stat(treatment, data, weights = rep(1, length(treatment)))
}
\arguments{
\item{data}{Confounders for which to compute standardized mean difference}

\item{weights}{Frequency weights for each data point}

\item{treament}{Binary treatment variable}
}
\value{
SMD statistic
}
\description{
Weighted SMD stat
}
