% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_smd.R
\name{weighted_smd_table}
\alias{weighted_smd_table}
\title{Weighted smd}
\usage{
weighted_smd_table(treatment, data, weights)
}
\arguments{
\item{treatment}{Binary treatment variable}

\item{data}{Confounders/data matrix over whose columns to compute standardized mean difference}

\item{weights}{Frequency weights for occurrences of data}
}
\value{
unweighted and weighted smd
}
\description{
Weighted smd
}
