% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_var}
\alias{weighted_var}
\title{Weighted variance}
\usage{
weighted_var(x, w)
}
\arguments{
\item{x}{A numeric vector for which to compute the weighted variance}

\item{w}{A numeric vector of the same length as x, with frequency weights (i.e. sum of weights is equal to number of occurrences)}
}
\value{
Returns the weighted variance
}
\description{
This function returns the weighted variance of a vector
}
