% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ks.R
\name{weighted_ks_stat}
\alias{weighted_ks_stat}
\title{Compute weighted 2-sample KS statistics}
\usage{
weighted_ks_stat(x, y, weights_x = NULL, weights_y = NULL)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}

\item{weights_x}{Weights for first sample}

\item{weights_y}{Weights for second sample}
}
\value{
Test statistic for weighted 2 samples
}
\description{
Compute weighted 2-sample KS statistics
}
